remove container script
