+++ Assets/Scripts/SS3D/Systems/Inventory/Containers/ClothesDisplayer.cs
@@ -92,7 +92,7 @@ namespace SS3D.Systems.Inventory.Containers
 		/// When the content of a container change, check if it should display or remove display of some clothes.
 		/// </summary>
 		[Server]
-        public void HandleContainerContentChanged(Container container, IEnumerable<Item> oldItems, IEnumerable<Item> newItems, ContainerChangeType type)
+        public void HandleContainerContentChanged(AttachedContainer container, IEnumerable<Item> oldItems, IEnumerable<Item> newItems, ContainerChangeType type)
         {
             // If it's not a cloth type container.
             // It'd be probably better to just create "cloth container" inheriting from container to easily test that.
