--- Assets/Scripts/SS3D/Systems/Inventory/Containers/Container.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cf2e1e93bb2b4f2eac8df7a07b518352
-timeCreated: 1666117887
