+++ Assets/Scripts/SS3D/Systems/Inventory/Containers/ContainerViewer.cs
@@ -68,7 +68,7 @@ namespace SS3D.Systems.Inventory.Containers
             }
 
             Hands hands = GetComponent<Hands>();
-            foreach (Entity observer in container.Container.ObservingPlayers)
+            foreach (Entity observer in container.ObservingPlayers)
             {
                 // checks if the container is already viewed by another entity
                 if (HasContainer(container) && observer != hands)
@@ -105,7 +105,6 @@ namespace SS3D.Systems.Inventory.Containers
         /// </summary>
         public void ShowContainerUI(AttachedContainer attachedContainer)
         {
-            attachedContainer.Container.AddObserver(GetComponent<Entity>());
             _displayedContainers.Add(attachedContainer);
             SetOpenState(attachedContainer.gameObject, true);
             NetworkConnection client = Owner;
@@ -120,7 +119,6 @@ namespace SS3D.Systems.Inventory.Containers
         /// </summary>
         public void CloseContainerUI(AttachedContainer container)
         {
-            container.Container.RemoveObserver(GetComponent<Entity>());
             if (_displayedContainers.Remove(container))
             {
                 Debug.Log("client call remove");
