+++ Assets/Scripts/SS3D/Systems/Inventory/Items/ItemSystem.cs
@@ -115,7 +115,7 @@ namespace SS3D.Systems.Inventory.Items
 
             Item itemInstance = Instantiate(itemPrefab, Vector3.zero, Quaternion.identity);
             ServerManager.Spawn(itemInstance.GameObject);
-            attachedContainer.Container.AddItem(itemInstance);
+            attachedContainer.AddItem(itemInstance);
 
             Punpun.Information(this, "Item {item} spawned in container {container}", Logs.ServerOnly, itemInstance.name, attachedContainer.ContainerName);
             return itemInstance;
