+++ Assets/Scripts/SS3D/Systems/Inventory/UI/SingleItemContainerSlot.cs
@@ -1,4 +1,4 @@
-using SS3D.Systems.Inventory.Containers;
+﻿using SS3D.Systems.Inventory.Containers;
 using SS3D.Systems.Inventory.Interfaces;
 using SS3D.Systems.Inventory.Items;
 using System.Collections.Generic;
@@ -56,12 +56,12 @@ namespace SS3D.Systems.Inventory.UI
         /// </summary>
         public override void OnItemDisplayDrop(ItemDisplay display)
         {
-            if (!_container.Container.Empty)
+            if (!_container.Empty)
             {
                 return;
             }
 
-            if (!_container.Container.CanContainItem(display.Item))
+            if (!_container.CanContainItem(display.Item))
             {
                 return;
             }
@@ -82,7 +82,7 @@ namespace SS3D.Systems.Inventory.UI
         /// </summary>
         private void UpdateDisplay()
         {
-            var item = _container.Container.Items.FirstOrDefault();
+            var item = _container.Items.FirstOrDefault();
             ItemDisplay.Item = item;
         }
 
@@ -98,14 +98,14 @@ namespace SS3D.Systems.Inventory.UI
 
             if (_container != null)
             {
-                _container.Container.OnContentsChanged -= ContainerContentsChanged;
+                _container.OnContentsChanged -= ContainerContentsChanged;
             }
 
-            newContainer.Container.OnContentsChanged += ContainerContentsChanged;
+            newContainer.OnContentsChanged += ContainerContentsChanged;
             _container = newContainer;
         }
 
-        private void ContainerContentsChanged(Container _, IEnumerable<Item> items, IEnumerable<Item> newItems, ContainerChangeType type)
+        private void ContainerContentsChanged(AttachedContainer _, IEnumerable<Item> items, IEnumerable<Item> newItems, ContainerChangeType type)
         {
             if (type != ContainerChangeType.Move)
             {
